# Issue
There are cases where customers have outstanding invoices in 10s of thousands and the reconciliation tool timesout on normal operations like fetching and allocating. This happens because, the `Invoice Limit` and `Payment Limit` are applied after pulling in all outstanding invoices and payments. This is inefficient. 

# Goal
To fix this permanently, a default limit of 50(which is still editable by customer) is applied and a search field is supplied for outstanding invoices and payments. 

Search uses Indexed columns - `voucher_no` and `against_voucher_no` in Payment Ledger. This along with limit variable, makes the DB query fast. Fetch(`Get Unreconciled Entries`) becomes a constant time operation(`O(n)` -> `O(50)`) for default settings.

# Changes
Limits are set to 50 as a default and are passed directly to the DB query. 2 new fields - `Filter on Invoice` and `Filter on Payment`, are added which pass its value to DB query. This way Vouchers are filtered at the query level.

<img width="1440" alt="Screenshot 2023-08-15 at 6 18 36 AM" src="https://github.com/frappe/erpnext/assets/3272205/a9f3b608-f959-40ea-af37-bca14466e463">

## before
```
In [13]: %time pr.get_unreconciled_entries()
CPU times: user 792 ms, sys: 29 ms, total: 821 ms
Wall time: 19.2 s

In [14]: len(pr.invoices)
Out[14]: 10000

In [15]: len(pr.payments)
Out[15]: 803
```

## after
```
In [8]: %time pr.get_unreconciled_entries()
CPU times: user 107 ms, sys: 5.64 ms, total: 113 ms
Wall time: 643 ms

In [9]: len(pr.payments)
Out[9]: 50

In [10]: len(pr.invoices)
Out[10]: 50
```


https://github.com/frappe/erpnext/assets/3272205/27c418df-79b5-424c-85ef-3949087a4ba8